pyproject.toml
[project]
name = "phillips-hue-interface"
version = "0.1.0"
description = "Clean, portable Philips Hue light control library"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.32.5",
    "python-dotenv>=1.0.0",
    "hue-entertainment-pykit>=0.9.4",
]

[project.scripts]
hue = "hue.cli:main"
